add crud_inventory page (/item)

+++ Assignment2/crud_app/assets/js/index.js
@@ -87,6 +87,31 @@ $("#add_event").submit(function(event){
 
 })
 
+$("#add_item").submit(function(item){
+	item.preventDefault();
+			var unindexed_array = $(this).serializeArray();
+			var data = {}
+
+			$.map(unindexed_array, function(n, i){
+					data[n['name']] = n['value']
+			})
+
+//		if (validateForm(data)) {  NEED TO VALIDATE LATER
+			var request = {
+					"url" : `http://localhost:3000/api/items`,
+					"method" : "POST",
+					"data" : data
+			}
+
+			$.ajax(request).done(function(response){
+					alert("Data Inserted Successfully!");
+			})
+
+//    }
+
+})
+
+
 
 $("#update_user").submit(function(event){
     event.preventDefault();
@@ -139,6 +164,31 @@ $("#update_event").submit(function(event){
 
 })
 
+$("#update_item").submit(function(item){
+    item.preventDefault();
+
+			var unindexed_array = $(this).serializeArray();
+			var data = {}
+
+			$.map(unindexed_array, function(n, i){
+					data[n['name']] = n['value']
+			})
+
+//		if (validateForm(data)) { NEED TO DO VALIDATION LATER
+			var request = {
+					"url" : `http://localhost:3000/api/item/${data.id}`,      //
+					"method" : "PUT",
+					"data" : data
+			}
+
+			$.ajax(request).done(function(response){
+					alert("Data Updated Successfully!");
+			})
+
+//    }
+
+})
+
 
 //Deletes a user and updates the page to show it
 if(window.location.pathname == "/"){
@@ -167,6 +217,26 @@ if(window.location.pathname == "/profile"){
 
         var request = {
             "url" : `http://localhost:3000/api/events/${id}`,
+            "method" : "DELETE"
+        }
+
+        if(confirm("Do you really want to delete this record?")){
+            $.ajax(request).done(function(response){
+                alert("Data Deleted Successfully!");
+                location.reload();
+            })
+        }
+
+    })
+}
+
+//Deletes a item and updates the page to show it
+if(window.location.pathname == "/items"){
+    $(".table tbody td a.delete").click(function(){
+        var id = $(this).attr("data-id")
+
+        var request = {
+            "url" : `http://localhost:3000/api/item/${id}`,
             "method" : "DELETE"
         }
